Cleaning up old template tests.
