+++ test/templates.js
@@ -1,260 +1,449 @@
 'use strict';
 
-let _ = require('lodash'); 
-let formioUtils = require('formio-utils');
-let request = require('supertest');
 let assert = require('assert');
-let DOCKER = process.env.DOCKER;
-let CUSTOMER = process.env.CUSTOMER;
+let _ = require('lodash');
+let formioUtils = require('formio-utils');
 
 module.exports = (app, template, hook) => {
-  describe('Template Tests', function() {
-    describe('Export Format', function() {
-      let exportTemplate;
+  describe('Template Imports', function() {
+    /**
+     * Util function to get the resource name, given an id.
+     *
+     * @param project
+     * @param id
+     * @returns {*}
+     */
+    let getResourceFromId = (project, id) => {
+      project = project || {};
+      project.forms = project.forms || {};
+      project.resources = project.resources || {};
 
-      before(function(done) {
-        request(app)
-          .get(hook.alter('url', '/export', template))
-          .set('x-jwt-token', template.users.admin.token)
-          .expect(200)
-          .expect('Content-Type', /json/)
-          .end((err, res) => {
+      let resourceName;
+      if (project.forms[id]) {
+        resourceName = project.forms[id].machineName;
+      }
+      else if (project.resources[id]) {
+        resourceName = project.resources[id].machineName;
+      }
+      else if (project.roles[id]) {
+        resourceName = project.roles[id].machineName
+      }
+
+      return resourceName;
+    };
+
+    /**
+     * Util function to compare the input template with the current resources in mongo.
+     *
+     * @param project
+     * @param input
+     * @param done
+     */
+    let checkTemplateRoles = (project, input, done) => {
+      input = input || {};
+
+      app.formio.resources.role.model.find({deleted: {$eq: null}}).then(roles => {
+        assert.equal(roles.length, Object.keys(input).length);
+
+        // If the input is empty, skip remaining checks.
+        if (Object.keys(input).length === 0) {
+          return done();
+        }
+
+        // Check that the template data doesnt contain any _id's or machineNames
+        Object.keys(input).forEach(machineName => {
+          let role = input[machineName];
+
+          assert.equal(role.hasOwnProperty('_id'), false);
+          assert.equal(role.hasOwnProperty('machineName'), false);
+        });
+
+        let given = {};
+
+        // Memoize the roles.
+        project.roles = {};
+        roles.forEach(role => {
+          role = role.toObject();
+
+          // Check that each role in mongo has an _id and a machineName.
+          assert.equal(role.hasOwnProperty('_id'), true);
+          assert.equal(role.hasOwnProperty('machineName'), true);
+
+          // Prepare the stored roles for comparison.
+          let machineName = role.machineName;
+          given[machineName] = _.omit(role, ['_id', '__v', 'created', 'deleted', 'modified', 'machineName']);
+
+          project.roles[machineName] = project.roles[role._id] = role;
+        });
+
+        assert.deepEqual(given, input);
+        done();
+      })
+      .catch(done);
+    };
+
+    /**
+     * Util function to compare the input template with the current resources in mongo.
+     *
+     * @param project
+     * @param type
+     * @param input
+     * @param done
+     */
+    let checkTemplateFormsAndResources = (project, type, input, done) => {
+      input = input || {};
+
+      app.formio.resources.form.model.find({type, deleted: {$eq: null}}).then(forms => {
+        assert.equal(forms.length, Object.keys(input).length);
+
+        if (Object.keys(input).length === 0) {
+          return done()
+        }
+
+        // Check that the template data doesnt contain any _id's or machineNames
+        Object.keys(input).forEach(machineName => {
+          let form = input[machineName];
+
+          assert.equal(form.hasOwnProperty('_id'), false);
+          assert.equal(form.hasOwnProperty('machineName'), false);
+        });
+
+        let given = {};
+
+        // Memoize the forms.
+        project[`${type}s`] = {};
+        forms.forEach(form => {
+          form = form.toObject();
+
+          // Check that each form in mongo has an _id and machineName.
+          assert.equal(form.hasOwnProperty('_id'), true);
+          assert.equal(form.hasOwnProperty('machineName'), true);
+
+          let machineName = form.machineName;
+          let tempForm = _.omit(form, ['_id', '__v', 'created', 'deleted', 'modified', 'machineName', 'owner']);
+
+          tempForm.access = tempForm.access.map(access => {
+            access.roles = access.roles.map(role => {
+              return project.roles[role.toString()].machineName;
+            });
+
+            return access;
+          });
+
+          tempForm.submissionAccess = tempForm.submissionAccess.map(access => {
+            access.roles = access.roles.map(role => {
+              return project.roles[role.toString()].machineName;
+            });
+
+            return access;
+          });
+
+          // Convert all resources to point to the resource name;
+          formioUtils.eachComponent(tempForm.components, (component) => {
+            if (component.hasOwnProperty('resource')) {
+              component.resource = project[`${type}s`][component.resource].name;
+            }
+          }, true);
+          given[machineName] = tempForm;
+
+          project[`${type}s`][form.machineName] = project[`${type}s`][form._id] = form;
+        });
+
+        assert.deepEqual(given, input);
+        done();
+      })
+      .catch(done);
+    };
+
+    /**
+     * Util function to compare the input template with the current resources in mongo.
+     *
+     * @param project
+     * @param input
+     * @param done
+     */
+    let checkTemplateActions = (project, input, done) => {
+      input = input || {};
+
+      app.formio.actions.model.find({deleted: {$eq: null}}).then(actions => {
+        assert.equal(actions.length, Object.keys(input).length);
+
+        if (Object.keys(input).length === 0) {
+          return done()
+        }
+
+        // Check that the template data doesnt contain any _id's or machineNames
+        Object.keys(input).forEach(machineName => {
+          let action = input[machineName];
+
+          assert.equal(action.hasOwnProperty('_id'), false);
+          assert.equal(action.hasOwnProperty('machineName'), false);
+        });
+
+        let given = {};
+
+        // Memoize the forms.
+        project.actions = {};
+        actions.forEach(action => {
+          action = action.toObject();
+
+          // Check that each action in mongo has an _id and machineName.
+          assert.equal(action.hasOwnProperty('_id'), true);
+          assert.equal(action.hasOwnProperty('machineName'), true);
+
+          // Prepare the stored actions for comparison.
+          let machineName = action.machineName;
+          let tempAction = _.omit(action, ['_id', '__v', 'created', 'deleted', 'modified', 'machineName']);
+          tempAction.form = getResourceFromId(project, tempAction.form);
+          if (_.has(tempAction, 'settings.resource')) {
+            tempAction.settings.resource = getResourceFromId(project, tempAction.settings.resource);
+          }
+          if (_.has(tempAction, 'settings.resources')) {
+            tempAction.settings.resources = tempAction.settings.resources.map(resource => {
+              return getResourceFromId(project, resource);
+            });
+          }
+          if (_.has(tempAction, 'settings.role')) {
+            tempAction.settings.role = project.roles[tempAction.settings.role].machineName;
+          }
+          given[machineName] = tempAction;
+
+          project.actions[machineName] = project.actions[action._id] = action;
+        });
+
+        assert.deepEqual(given, input);
+        done();
+      })
+      .catch(done);
+    };
+
+    describe('Empty Template', function() {
+      let emptyTemplate = require('../src/templates/empty.json');
+      let _template = _.cloneDeep(emptyTemplate);
+
+      describe('Import', function() {
+        let project = {};
+
+        it('Should be able to bootstrap the empty template', function(done) {
+          app.formio.template.import(_template, (err) => {
             if (err) {
               return done(err);
             }
 
-            exportTemplate = res.body;
-            assert.deepEqual(
-              Object.keys(exportTemplate),
-              ['title', 'version', 'description', 'name', 'plan', 'roles', 'forms', 'actions', 'resources']
-            );
+            done();
+          });
+        });
 
+        it('All the roles should be imported', function(done) {
+          checkTemplateRoles(project, emptyTemplate.roles, done);
+        });
+
+        it('No forms should exist', function(done) {
+          checkTemplateFormsAndResources(project, 'form', emptyTemplate.forms, done);
+        });
+
+        it('No resources should exist', function(done) {
+          checkTemplateFormsAndResources(project, 'resource', emptyTemplate.resources, done);
+        });
+
+        it('No actions should exist', function(done) {
+          checkTemplateActions(project, emptyTemplate.actions, done);
+        });
+      });
+
+      describe('Export', function() {
+        let project = {};
+        let exportData = {};
+        
+        it('Should be able to export project data', function(done) {
+          app.formio.template.export(emptyTemplate, (err, data) => {
+            if (err) {
+              return done(err);
+            }
+            
+            exportData = data;
             return done();
           });
-      });
-
-      it('An export should contain the export title', function() {
-        assert.equal(
-          hook.alter('exportTitle', 'Export', template),
-          'Export'
-        );
-      });
-
-      it('An export should contain the current export version', function() {
-        assert.equal(
-          exportTemplate.version,
-          '2.0.0'
-        );
-      });
-
-      it('An export should contain the description', function() {
-        assert.equal(
-          hook.alter('exportDescription', '', template),
-          ''
-        );
-      });
-
-      it('An export should contain the export name', function() {
-        assert.equal(
-          hook.alter('exportName', 'export', template),
-          'export'
-        );
-      });
-
-      it('An export should contain the export plan', function() {
-        assert.equal(
-          hook.alter('exportPlan', 'community', template),
-          'community'
-        );
-      });
-
-      // Skip the remaining tests if this is a docker or customer test.
-      let project = {};
-      if (DOCKER || CUSTOMER) {
-        return;
-      }
-      it('Grab all the project resources', function(done) {
-        let roleP = app.formio.resources.role.model.find({deleted: {$eq: null}}).then(roles => {
-          project.roles = {};
-          roles.forEach(role => {
-            role = role.toObject();
-
-            project.roles[role._id] = role;
-          });
         });
 
-        let formP = app.formio.resources.form.model.find({type: 'form', deleted: {$eq: null}}).then(forms => {
-          project.forms = {};
-          forms.forEach(form => {
-            form = form.toObject();
-
-            project.forms[form._id] = form;
-          });
+        it('An export should contain the export title', function() {
+          assert.equal(
+            hook.alter('exportTitle', 'Export', exportData),
+            'Export'
+          );
         });
 
-        let resourceP = app.formio.resources.form.model.find({type: 'resource', deleted: {$eq: null}}).then(resources => {
-          project.resources = {};
-          resources.forEach(resource => {
-            resource = resource.toObject();
-
-            project.resources[resource._id] = resource;
-          });
+        it('An export should contain the current export version', function() {
+          assert.equal(
+            exportData.version,
+            '2.0.0'
+          );
         });
 
-        let actionP = app.formio.actions.model.find({deleted: {$eq: null}}).then(actions => {
-          project.actions = {};
-          actions.forEach(action => {
-            action = action.toObject();
-
-            project.actions[action._id] = action;
-          });
+        it('An export should contain the description', function() {
+          assert.equal(
+            hook.alter('exportDescription', '', exportData),
+            ''
+          );
         });
 
-        Promise.all([roleP, formP, resourceP, actionP])
-        .then(() => done())
-        .catch(done);
+        it('An export should contain the export name', function() {
+          assert.equal(
+            hook.alter('exportName', 'export', exportData),
+            'export'
+          );
+        });
+
+        it('An export should contain the export plan', function() {
+          assert.equal(
+            hook.alter('exportPlan', 'community', exportData),
+            'community'
+          );
+        });
+
+        it('The empty template should export all its roles', function(done) {
+          checkTemplateRoles(project, exportData.roles, done);
+        });
+
+        it('The empty template should not export any forms', function(done) {
+          checkTemplateFormsAndResources(project, 'form', exportData.forms, done);
+        });
+
+        it('The empty template should not export any resources', function(done) {
+          checkTemplateFormsAndResources(project, 'resource', exportData.resources, done);
+        });
+
+        it('The empty template should not export any actions', function(done) {
+          checkTemplateActions(project, exportData.actions, done);
+        });
+
+        it('An export should match an import', function() {
+          assert.equal(exportData.version, '2.0.0');
+          assert.deepEqual(_.omit(exportData, ['plan', 'version']), _.omit(emptyTemplate, ['plan', 'version']));
+        });
       });
 
-      it('An export should contain the project roles', function(done) {
-        app.formio.resources.role.model.find({deleted: {$eq: null}}).then(roles => {
-          let expectedRoles = exportTemplate.roles;
-          let responseRoles = {};
-
-          roles.forEach(role => {
-            let tempRole = _.omit(role.toObject(), ['_id', '__v', 'created', 'deleted', 'modified']);
-            let machineName = tempRole.machineName;
-            delete tempRole.machineName;
-
-            responseRoles[machineName] = tempRole;
-          });
-
-          assert.deepEqual(responseRoles, expectedRoles);
-          return done();
-        })
-        .catch(done);
+      before(function(done) {
+        template.clearData(done);
       });
 
-      it('An export should contain the project forms', function(done) {
-        app.formio.resources.form.model.find({type: 'form', deleted: {$eq: null}}).then(forms => {
-          let expectedForms = exportTemplate.forms;
-          let responseForms = {};
-          forms.map(form => {
-            let machineName = form.toObject().machineName;
-            let tempForm = _.omit(form.toObject(), ['_id', '__v', 'created', 'deleted', 'modified', 'machineName', 'owner']);
-
-            tempForm.access = tempForm.access.map(access => {
-              access.roles = access.roles.map(role => {
-                return project.roles[role.toString()].machineName;
-              });
-
-              return access;
-            });
-
-            tempForm.submissionAccess = tempForm.submissionAccess.map(access => {
-              access.roles = access.roles.map(role => {
-                return project.roles[role.toString()].machineName;
-              });
-
-              return access;
-            });
-
-            // Convert all resources to point to the resource name;
-            formioUtils.eachComponent(tempForm.components, (component) => {
-              if (component.hasOwnProperty('resource')) {
-                component.resource = project.resources[component.resource].name;
-              }
-            }, true);
-
-            responseForms[machineName] = tempForm;
-          });
-
-          assert.deepEqual(responseForms, expectedForms);
-          return done();
-        })
-        .catch(done);
+      after(function(done) {
+        template.clearData(done);
       });
+    });
 
-      it('An export should contain the project resources', function(done) {
-        app.formio.resources.form.model.find({type: 'resource', deleted: {$eq: null}}).then(resources => {
-          let expectedResources = exportTemplate.resources;
-          let responseResources = {};
-          resources.map(resource => {
-            let machineName = resource.toObject().machineName;
-            let tempResource = _.omit(resource.toObject(), ['_id', '__v', 'created', 'deleted', 'modified', 'machineName', 'owner']);
+    describe('Default Template', function() {
+      let defaultTemplate = require('../src/templates/default.json');
+      let _template = _.cloneDeep(defaultTemplate);
 
-            tempResource.access = tempResource.access.map(access => {
-              access.roles = access.roles.map(role => {
-                return project.roles[role.toString()].machineName;
-              });
+      describe('Import', function() {
+        let project = {};
 
-              return access;
-            });
-
-            tempResource.submissionAccess = tempResource.submissionAccess.map(access => {
-              access.roles = access.roles.map(role => {
-                return project.roles[role.toString()].machineName;
-              });
-
-              return access;
-            });
-
-            // Convert all resources to point to the resource name;
-            formioUtils.eachComponent(tempResource.components, (component) => {
-              if (component.hasOwnProperty('resource')) {
-                component.resource = project.resources[component.resource].name;
-              }
-            }, true);
-
-            responseResources[machineName] = tempResource;
-          });
-
-          assert.deepEqual(responseResources, expectedResources);
-          return done();
-        })
-        .catch(done);
-      });
-
-      it('An export should contain the project actions', function(done) {
-        let getResourceFromId = (id) => {
-          let resourceName;
-
-          if (project.forms[id]) {
-            resourceName = project.forms[id].name;
-          }
-          else if (project.resources[id]) {
-            resourceName = project.resources[id].name;
-          }
-
-          return resourceName;
-        };
-
-        app.formio.actions.model.find({deleted: {$eq: null}}).then(actions => {
-          let expectedActions = exportTemplate.actions;
-          let responseActions = {};
-          actions.map(action => {
-            let machineName = action.toObject().machineName;
-            let tempAction = _.omit(action.toObject(), ['_id', '__v', 'created', 'deleted', 'modified', 'machineName']);
-
-            tempAction.form = getResourceFromId(tempAction.form);
-            if (_.has(tempAction, 'settings.resource')) {
-              tempAction.settings.resource = getResourceFromId(tempAction.settings.resource);
-            }
-            if (_.has(tempAction, 'settings.resources')) {
-              tempAction.settings.resources = tempAction.settings.resources.map(resource => {
-                return getResourceFromId(resource);
-              });
-            }
-            if (_.has(tempAction, 'settings.role')) {
-              tempAction.settings.role = project.roles[tempAction.settings.role].machineName;
+        it('Should be able to bootstrap the default template', function(done) {
+          app.formio.template.import(_template, (err) => {
+            if (err) {
+              return done(err);
             }
 
-            responseActions[machineName] = tempAction;
+            done();
           });
+        });
 
-          assert.deepEqual(responseActions, expectedActions);
-          return done();
-        })
-        .catch(done);
+        it('All the roles should be imported', function(done) {
+          checkTemplateRoles(project, defaultTemplate.roles, done);
+        });
+
+        it('All the forms should be imported', function(done) {
+          checkTemplateFormsAndResources(project, 'form', defaultTemplate.forms, done);
+        });
+
+        it('All the resources should be imported', function(done) {
+          checkTemplateFormsAndResources(project, 'resource', defaultTemplate.resources, done);
+        });
+
+        it('All the actions should be imported', function(done) {
+          checkTemplateActions(project, defaultTemplate.actions, done);
+        });
+      });
+
+      describe('Export', function() {
+        let project = {};
+        let exportData = {};
+
+        it('Should be able to export project data', function(done) {
+          app.formio.template.export(defaultTemplate, (err, data) => {
+            if (err) {
+              return done(err);
+            }
+
+            exportData = data;
+            return done();
+          });
+        });
+
+        it('An export should contain the export title', function() {
+          assert.equal(
+            hook.alter('exportTitle', 'Export', exportData),
+            'Export'
+          );
+        });
+
+        it('An export should contain the current export version', function() {
+          assert.equal(
+            exportData.version,
+            '2.0.0'
+          );
+        });
+
+        it('An export should contain the description', function() {
+          assert.equal(
+            hook.alter('exportDescription', '', exportData),
+            ''
+          );
+        });
+
+        it('An export should contain the export name', function() {
+          assert.equal(
+            hook.alter('exportName', 'export', exportData),
+            'export'
+          );
+        });
+
+        it('An export should contain the export plan', function() {
+          assert.equal(
+            hook.alter('exportPlan', 'community', exportData),
+            'community'
+          );
+        });
+
+        it('The Default template should export all its roles', function(done) {
+          checkTemplateRoles(project, exportData.roles, done);
+        });
+
+        it('The Default template should not export any forms', function(done) {
+          checkTemplateFormsAndResources(project, 'form', exportData.forms, done);
+        });
+
+        it('The Default template should not export any resources', function(done) {
+          checkTemplateFormsAndResources(project, 'resource', exportData.resources, done);
+        });
+
+        it('The Default template should not export any actions', function(done) {
+          checkTemplateActions(project, exportData.actions, done);
+        });
+
+        it('An export should match an import', function() {
+          assert.equal(exportData.version, '2.0.0');
+          assert.deepEqual(_.omit(exportData, ['plan', 'version']), _.omit(defaultTemplate, ['plan', 'version']));
+        });
+      });
+
+      before(function(done) {
+        template.clearData(done);
+      });
+
+      after(function(done) {
+        template.clearData(done);
       });
     });
   });
