+++ test/test.js
@@ -54,11 +54,10 @@ describe('Initialization', function() {
   });
 
   after(function() {
-    require('./template-import')(app, template, hook);
+    require('./templates')(app, template, hook);
     require('./bootstrap')(app, template, hook);
     require('./unit')(app, template, hook);
     require('./auth')(app, template, hook);
-    require('./templates')(app, template, hook);
     require('./roles')(app, template, hook);
     require('./form')(app, template, hook);
     require('./resource')(app, template, hook);
